pyproject.toml
[project]
name = "praxis-deid"
version = "0.1.0"
description = "Practice-side de-identification tool for Praxis. Produces HIPAA Safe Harbor-compliant aggregates from raw practice management data, before anything crosses to Praxis cloud."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Praxis", email = "engineering@praxishealth.ai" }]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: Office/Business :: Medical",
]
dependencies = [
  "pyyaml>=6.0",
  "requests>=2.31",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "hypothesis>=6.0",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-PyYAML",
  "types-requests",
]

[project.scripts]
praxis-deid = "praxis_deid.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["praxis_deid*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S"]
ignore = ["S101"] # asserts in tests

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
